a big dump (tracker)
Canonical coordinates calculated in bunch generation correctly.
Stored as plain members in TRKParticle for now.  Not tidy yet.

Added TRKKicker element, with necessary edits to existing strategies,
base classes, and TRKFactory.

Added many integrators to TRKMaps from which Track methods have been
implemented in TRKDefaultStrategy.

Added some hacks to try to get the output writing to be correct enough
for rebdsimOptics to stop segfaulting, which explains some mess in
TRKBunch and BDSOutputROOTEventSampler (especially).  Will be
bettered.  Added S position to this end for example to TRKSampler,
which may be not particularly clean but got be closer in my attempt to
tick all the boxes for rebdsimOptics to not segfault.


Former-commit-id: 6b2aafabee2fd6bd456f2ad47541aa6292123916